+++ tensorflow_io/core/python/ops/v0/hdf5_io_tensor_ops.py
@@ -21,7 +21,7 @@ import uuid
 
 import tensorflow as tf
 from tensorflow_io.core.python.ops import core_ops
-from tensorflow_io.core.python.ops import io_tensor_ops
+from tensorflow_io.core.python.ops.v0 import io_tensor_ops
 
 class HDF5IOTensor(io_tensor_ops._CollectionIOTensor): # pylint: disable=protected-access
   """HDF5IOTensor"""
