+++ tensorflow_io/core/python/ops/v0/tiff_io_tensor_ops.py
@@ -18,7 +18,7 @@ from __future__ import division
 from __future__ import print_function
 
 import tensorflow as tf
-from tensorflow_io.core.python.ops import io_tensor_ops
+from tensorflow_io.core.python.ops.v0 import io_tensor_ops
 from tensorflow_io.core.python.ops import core_ops
 
 class TIFFIOTensor(io_tensor_ops._CollectionIOTensor): # pylint: disable=protected-access
