+++ tensorflow_io/prometheus/python/ops/prometheus_ops.py
@@ -20,8 +20,8 @@ from __future__ import print_function
 import warnings
 
 import tensorflow as tf
-from tensorflow_io.core.python.ops import data_ops
-from tensorflow_io.core.python.ops import core_golang_ops
+from tensorflow_io.core.python.ops import golang_ops
+from tensorflow_io.core.python.ops.v0 import data_ops
 
 warnings.warn(
     "The tensorflow_io.prometheus.PrometheusDataset is "
@@ -32,7 +32,7 @@ warnings.warn(
 
 def read_prometheus(endpoint, query):
   """read_prometheus"""
-  return core_golang_ops.io_read_prometheus(endpoint, query)
+  return golang_ops.io_read_prometheus(endpoint, query)
 
 class PrometheusDataset(data_ops.BaseDataset):
   """A Prometheus Dataset"""
